exp awarding / tweaks
- changed Exp to a float
- Taming a slime gives exp = its rarity * 1.5
- Splicing a slime gives exp = its rarity
- crops now have an expYeild value that is awarded when harvested
- EXP bar will now update after leveling up (so it wont stay at full until more exp is earned)

+++ Slime Garden/Assets/Scripts/DataPersistence/GameData.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class GameData
 {
     public int level;
-    public int experience;
+    public float experience;
     public int money;
     public int habitatTier;
     public Vector2Int borderFenceStyle;
